Added RNACentral reference DB; added a reference to ids mapping in the DB trait, removed from parameters
