+++ src/main/scala/metagenomica/parameters.scala
@@ -29,8 +29,7 @@ trait AnyMG7Parameters {
   // TODO: would be nice to have Nat here
   val chunkSize: Int
 
-  val referenceDB: bundles.AnyBlastReferenceDB
-  val referenceMap: bundles.AnyReferenceMap
+  val referenceDB: bundles.AnyReferenceDB
 }
 
 abstract class MG7Parameters[
@@ -38,8 +37,7 @@ abstract class MG7Parameters[
 ](val readsLength: illumina.Length,
   val blastOutRec: BR,
   val chunkSize: Int = 5,
-  val referenceDB: bundles.AnyBlastReferenceDB,
-  val referenceMap: bundles.AnyReferenceMap
+  val referenceDB: bundles.AnyReferenceDB
 // )(implicit
   // TODO: add a check for minimal set of properties in the record (like bitscore and sgi)
 ) extends AnyMG7Parameters {
